+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Unreleased
 
 * Feat: Add Culture Context (#491)
+* Remove `Sentry.currentHub` (#490)
 
 # 5.1.0